removed unnecessary directory, added @ to php
You don't need to specify `./vendor/bin/`, composer is smart enough to find the binaries. Also, removing the directories makes the scripts usable on Windows too.
Also, added `@` before `php` in the `phpunit-coverage` script. This makes sure that the PHP version to run the coverage test is the same as composer is currently running on.